Update libc version and Cargo version for publishin

r? @mbrubeck

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.png" height=40 alt="Review on Reviewable"/>](https://reviewable.io/reviews/servo/gleam/52)
<!-- Reviewable:end -->

+++ src/gl.rs
@@ -30,7 +30,7 @@ pub fn buffer_data<T>(target: GLenum, data: &[T], usage: GLenum) {
 }
 
 #[inline]
-pub fn buffer_sub_data<T>(target: GLenum, offset: i64, data: &[T]) {
+pub fn buffer_sub_data<T>(target: GLenum, offset: isize, data: &[T]) {
     unsafe {
         ffi::BufferSubData(target,
                            offset,
